refactor(models): change ID fields from int to long
Update the `goalId` and `taskId` fields in the `Goals` and `Task` models from `int` to `long` to accommodate larger ID values and prevent potential integer overflow issues.

+++ goals-management/src/main/java/com/example/goals_management/models/Goals.java
@@ -19,7 +19,7 @@ public class Goals {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "goal_id")
-    private int goalId;
+    private long goalId;
 
     private String description;
